team-2-project/requirements.txt
boto3==1.17.28
botocore==1.20.28
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-dotenv==0.15.0
s3transfer==0.3.4
six==1.15.0
urllib3==1.26.4
Werkzeug==1.0.1
